Pass services  which are wip to keep backend unchanged

+++ back-end/app/controllers/organizations_controller.rb
@@ -122,6 +122,7 @@ class OrganizationsController < ApplicationController
         services_to_add.each do |service_id|
           OrgService.create(organization: @organization, service_id: service_id)
         end
+      elsif params[:organization][:service_ids].empty?
         # If no services are selected, remove all existing connections and their associated requests
         @organization.org_services.each do |org_service|
           org_service.requests.destroy_all
